requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.4.0
dashscope>=1.20.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
fastmcp>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
